WIP New inventory model first draft
(not definitive)

Keep only the necessary parts, formalize more explicitly model

Getters/Setters/Constructors not generated yet.
