+++ hawkular-inventory-parent/hawkular-inventory-api/src/main/java/org/hawkular/inventory/api/oldmodel/StructuredData.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.inventory.api.model;
+package org.hawkular.inventory.api.oldmodel;
 
 import java.io.IOException;
 import java.io.Serializable;
